Move files to more appropriate folder; Add new pipeline hook; Move providers to shared
