+++ application/ui/src/features/annotator/annotator-canvas.tsx
@@ -7,10 +7,10 @@ import { API_BASE_URL } from 'src/api/client';
 import { DatasetItem } from 'src/constants/shared-types';
 
 import { ZoomTransform } from '../../components/zoom/zoom-transform';
-import { useAnnotationActions } from './annotation-actions-provider.component';
-import { useAnnotationVisibility } from './annotation-visibility-provider.component';
+import { useAnnotationActions } from '../../shared/annotator/annotation-actions-provider.component';
+import { useAnnotationVisibility } from '../../shared/annotator/annotation-visibility-provider.component';
+import { useSelectedAnnotations } from '../../shared/annotator/select-annotation-provider.component';
 import { Annotations } from './annotations/annotations.component';
-import { useSelectedAnnotations } from './select-annotation-provider.component';
 import { ToolManager } from './tools/tool-manager.component';
 
 const getImageUrl = (projectId: string, itemId: string) => {
