+++ application/ui/src/features/annotator/tools/segment-anything-tool/segment-anything-tool.component.tsx
@@ -6,10 +6,10 @@ import { PointerEvent, useEffect, useRef, useState } from 'react';
 import { clampPointBetweenImage } from '@geti/smart-tools/utils';
 
 import { useZoom } from '../../../../components/zoom/zoom.provider';
+import { useAnnotator } from '../../../../shared/annotator/annotator-provider.component';
 import { AnnotationShapeWithLabels } from '../../annotations/annotation-shape-with-labels.component';
 import { MaskAnnotations } from '../../annotations/mask-annotations.component';
 import { AnnotatorLoading } from '../../annotator-loading.component';
-import { useAnnotator } from '../../annotator-provider.component';
 import { Annotation, Shape } from '../../types';
 import { SvgToolCanvas } from '../svg-tool-canvas.component';
 import { getRelativePoint, removeOffLimitPoints } from '../utils';
