+++ application/ui/src/features/annotator/tools/segment-anything-tool/use-decoding-query.hook.ts
@@ -3,8 +3,8 @@
 
 import { queryOptions, useMutation, useQuery } from '@tanstack/react-query';
 
-import { useAnnotationActions } from '../../annotation-actions-provider.component';
-import { useAnnotator } from '../../annotator-provider.component';
+import { useAnnotationActions } from '../../../../shared/annotator/annotation-actions-provider.component';
+import { useAnnotator } from '../../../../shared/annotator/annotator-provider.component';
 import { Shape } from '../../types';
 import { removeOffLimitPoints } from '../utils';
 import { InteractiveAnnotationPoint } from './segment-anything.interface';
