+++ application/ui/src/features/annotator/tools/tool-manager.component.tsx
@@ -1,7 +1,7 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { useAnnotator } from '../annotator-provider.component';
+import { useAnnotator } from '../../../shared/annotator/annotator-provider.component';
 import { BoundingBoxTool } from './bounding-box-tool/bounding-box-tool.component';
 import { SegmentAnythingTool } from './segment-anything-tool/segment-anything-tool.component';
