+++ application/ui/src/features/dataset/gallery/toolbar/toolbar.component.tsx
@@ -6,11 +6,11 @@ import { useQueryClient } from '@tanstack/react-query';
 import { useProjectIdentifier } from 'hooks/use-project-identifier.hook';
 import { DatasetItem } from 'src/constants/shared-types';
 
-import { $api } from '../../../api/client';
-import { AddMediaButton } from '../../../components/add-media-button/add-media-button.component';
-import { CheckboxInput } from '../../../components/checkbox-input/checkbox-input.component';
-import { DeleteMediaItem } from '../gallery/delete-media-item/delete-media-item.component';
-import { useSelectedData } from '../selected-data-provider.component';
+import { $api } from '../../../../api/client';
+import { AddMediaButton } from '../../../../components/add-media-button/add-media-button.component';
+import { CheckboxInput } from '../../../../components/checkbox-input/checkbox-input.component';
+import { DeleteMediaItem } from '../../gallery/delete-media-item/delete-media-item.component';
+import { useSelectedData } from '../../selected-data-provider.component';
 import { toggleMultipleSelection, updateSelectedKeysTo } from './util';
 
 type ToolbarProps = {
