+++ application/ui/src/features/dataset/media-preview/primary-toolbar/settings/settings.component.tsx
@@ -3,7 +3,7 @@
 
 import { Divider } from '@geti/ui';
 import { Adjustments, Invisible, LabelGroup, Visible } from '@geti/ui/icons';
-import { useAnnotationVisibility } from 'src/features/annotator/annotation-visibility-provider.component';
+import { useAnnotationVisibility } from 'src/shared/annotator/annotation-visibility-provider.component';
 
 import { IconWrapper } from '../../../../../components/icon-wrapper/icon-wrapper.component';
 import { ZoomFitScreen } from '../zoom/zoom-fit-screen.component';
