+++ application/ui/src/features/dataset/media-preview/secondary-toolbar/use-secondary-toolbar-state.hook.ts
@@ -4,9 +4,9 @@
 import { Key } from 'react';
 
 import { Label } from 'src/constants/shared-types';
-import { useAnnotationActions } from 'src/features/annotator/annotation-actions-provider.component';
 import { useProjectLabels } from 'src/features/annotator/hooks/use-project-labels.hook';
-import { useSelectedAnnotations } from 'src/features/annotator/select-annotation-provider.component';
+import { useAnnotationActions } from 'src/shared/annotator/annotation-actions-provider.component';
+import { useSelectedAnnotations } from 'src/shared/annotator/select-annotation-provider.component';
 
 export const useSecondaryToolbarState = () => {
     const { selectedAnnotations } = useSelectedAnnotations();
