+++ application/ui/src/features/inference/aside/data-collection.component.tsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from 'react';
 import { Divider, Flex, Heading, Slider, Switch, Text } from '@geti/ui';
 import { usePatchPipeline, usePipeline } from 'hooks/api/pipeline.hook';
 import { useIsPipelineConfigured } from 'hooks/use-is-pipeline-configured.hook';
-import { useProjectIdentifier } from 'src/hooks/use-project-identifier.hook';
+import { useProjectIdentifier } from 'hooks/use-project-identifier.hook';
 
 const DEFAULTS = {
     RATE: 12,
