+++ application/ui/src/features/inference/header/input-output-setup.component.tsx
