+++ application/ui/src/shared/annotator/annotation-visibility-provider.component.tsx
