+++ application/ui/src/shared/annotator/select-annotation-provider.component.tsx
